# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank REST API
  description: API для банковского приложения
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Generated server url
...
paths:
...
  /admin/users:
    get:
      tags:
        - User Management (Admin)
      summary: Получить список обычных пользователей
      description: Возвращает список пользователей с ролью USER с пагинацией и фильтрацией
        по имени
      operationId: getAllUser
      parameters:
        - name: username
          in: query
          description: Фильтр по имени пользователя
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Номер страницы (min = 1)
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: size
          in: query
          description: Размер страницы (min = 1)
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        "200":
          description: Список пользователей
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    post:
      tags:
        - User Management (Admin)
      summary: Создать нового пользователя
      description: Создаёт пользователя с ролью USER
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
        required: true
      responses:
        "201":
          description: Пользователь успешно создан
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /admin/users/admin:
    post:
      tags:
        - User Management (Admin)
      summary: Создать администратора
      description: Создаёт нового пользователя с ролью ADMIN
      operationId: createAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
        required: true
      responses:
        "201":
          description: Администратор успешно создан
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /admin/users/{id}:
    get:
      tags:
        - User Management (Admin)
      summary: Получить информацию о пользователе по ID
      description: Возвращает данные пользователя по UUID
      operationId: getOne
      parameters:
        - name: id
          in: path
          description: UUID пользователя
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Успешный ответ
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      tags:
        - User Management (Admin)
      summary: Удалить пользователя
      description: Логическое или физическое удаление пользователя по UUID
      operationId: delete
      parameters:
        - name: id
          in: path
          description: UUID пользователя
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Пользователь успешно удалён
    patch:
      tags:
        - User Management (Admin)
      summary: Обновить данные пользователя
      description: Частично обновляет данные пользователя по UUID
      operationId: update
      parameters:
        - name: id
          in: path
          description: UUID пользователя
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
        required: true
      responses:
        "200":
          description: Данные успешно обновлены
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /admin/users/admins:
    get:
      tags:
        - User Management (Admin)
      summary: Получить список администраторов
      description: Возвращает список пользователей с ролью ADMIN с пагинацией и фильтрацией
        по имени
      operationId: getAllAdmin
      parameters:
        - name: username
          in: query
          description: Фильтр по имени администратора
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Номер страницы (min = 1)
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: size
          in: query
          description: Размер страницы (min = 1)
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        "200":
          description: Список администраторов
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
components:
  schemas:
...
    UserCreateRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        repeat_password:
          type: string
          minLength: 1
      required:
        - password
        - repeat_password
        - username
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
      required:
        - id
        - username
...
    UserUpdateRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
        repeat_password:
          type: string
      required:
        - username
...